Update to Rails 2.1.2 (broken)
Unfortunately this is a broken commit (not bisectable) because Rails
still has bugs which prevent the specs from running. Nevertheless I
wanted the update and the patches that fix the bugs to appear as
separate commits so that the fixes can be easily found and re-applied
in subsequent releases if they continue to have the same bugs.

Signed-off-by: Wincent Colaiuta <win@wincent.com>