Kludge fix for Rails 2.2.0 breakage (broken extensions to ActiveRecord::Base)
Not sure what the problem is here but turning off class caching fixes
the issue. This is an undesirable kludge because it has a performance
penalty (classes are reloaded on every request).

My efforts to diagnose this problem have been unfruitful.

Logging statements show modules being included at startup time, and
methods like "acts_as_searchable" are present in ActiveRecord::Base.
But when the model files (eg. "article.rb") are evaluated, the
"acts_as_searchable" method is not present and the module is not
shown as having been included.

The order of the statements suggest that the module was indeed
included before the model file was evaluated, so either a buffering
glitch is causing the statements to be issued out of order, or
Rails is running in multi-threaded mode despite the fact that it
hasn't been configured to do so, or Rails is actually removing the
methods after inclusion (although this latter case seems unlikely).

I noticed that the bug only manifests itself in the test and production
environments, but not in the development one, so the kludge fix here
is to turn "cache_classes" off in the failing environments, bringing
them into line with the development environment.

Will continue to look for the real cause and the best solution.

Note that changing the order of the "require" statements in the
environment.rb file is not an option (ActiveRecord::Base is not even
defined yet inside the Rails::Initializer.run block).

Signed-off-by: Wincent Colaiuta <win@wincent.com>