More 2.2.0 breakage: special handling of routing for non-new records
Looks like I was shielded from this prior to 2.2.0 because the exact
same routing method worked for both new and non-new records.

Now I need to special case these.

This highlights what was really an abuse of my helper methods up until
now and they should really be refactored in the future. Specifically,
I am using "polymorphic_comment_path" only for "show" actions (and
automatically adding an anchor), when really this is the same URL that
should be used for "edit" actions as well (obviously, without the
anchor). Then I am abusing the "polymorphic_comments_path" helper for
"create" actions (which is appropriate) as well as for "edit" actions
(which is not appropriate).

Signed-off-by: Wincent Colaiuta <win@wincent.com>