Better fix for Rails 2.2.0 model preloading breakage
This was previously wallpapered over in commit a5d0bdc, but here is the
proper fix for the problem.

Rails 2.2.0 (and 2.2.1, 2.2.2), preload model classes at the end of the
intializer block. This means that any require statements that appear
after the intializer block won't have been evaluated, causing
exceptions to be raised and derailing the application.

This preloading only occurs when config.cache_classes is set to true.

The require statements can't be moved before or inside the initializer
block because their dependencies aren't available at that point.

The only solution is to use a separate initializer file. This will
cause them to be evaluated after the rest of the initializer block has
been executed (and all their dependencies are available to them), but
before the model classes are preloaded and control subsequently returns
from the initializer block.

Signed-off-by: Wincent Colaiuta <win@wincent.com>