Revert cdd9125: move application_controller.rb back to application.rb
Well, finally discovered that the famous Rails commit changing the name
of application.rb was actually made after 2.2.2 was released on what
looks to be a totally different branch (2.2.2 is not an ancestor of the
commit).

So back this out and hold tight until it makes it into a released
version.

Signed-off-by: Wincent Colaiuta <win@wincent.com>