Rails 2.2 fix: repair Haml "ugly" mode breakage
Yet another load ordering issue: Haml's ugly mode must now be set up
from an initializer file because doing it in environment.rb is now
to late (no runtime errors, but it appears the Haml has already been
set up by that point and the templates most likely precompiled).

The problem only appears in production, where class caching is turned
on. Hard to add a regression test for this one because class caching
is turned off in the test environment at the moment.

Signed-off-by: Wincent Colaiuta <win@wincent.com>