Change various TEXT column types to MEDIUMTEXT
Trying to post in a lengthy console session transcript in this ticket:

  http://rails.wincent.com/issues/1189

I noticed that the content was truncated unexpectedly. Turns out that I
was bumping into the 64K limit on the MySQL TEXT column type.

Here we add migrations to change several such columns to the MEDIUMTEXT
type, which can hold up to 16M. The columns in question are: comment
body, article body, issue description, post body (but not post excerpt),
and topic body.

TEXT columns left untouched were: product description and session data.

In practice, we'll never want to hold 16M of data in any fields, but
POST data will be clipped to 1M anyway by nginx due to the
client_max_body_size limit.

For security reasons, I'll be adding validations to prevent really huge
POSTs from filling up the disk too quickly, although that might be just
splitting hairs seeing as if an attacker can fill up the disk with
1,000 POSTs of 1M POSTs then he could probably just as easily
do it with 1,000,000 POSTs of 1K. (Of course, causing that much load on
the server would probably come to my notice long before the disk filled
up anyway.)

In any case, adding validations is a good idea from a UI perspective as
it would have helped me when I made the post in the ticket referenced
above and couldn't work out why the output was truncated.

See also:

  http://rails.wincent.com/issues/1190

Signed-off-by: Wincent Colaiuta <win@wincent.com>