Revise limit in MEDIUMTEXT migrations
Turns out that 16,777,215 doesn't work as we thought it did. If you use
that value then you wind up with LONGTEXT columns, not MEDIUMTEXT ones.
I am guessing that this is because the tables are set up to use UTF-8
encoding, so MySQL bumps the size up to 16,777,215 (characters) times
by 3 (maximum bytes per character) to arrive at a total size which
pushes into the LONGTEXT size range.

Problem caught because Rails doesn't properly handle LONGTEXT columns,
which in turn caused a bunch of specs to break.

Seeing as we haven't deployed this yet, just revise the migrations
themselves rather than adding new ones on top.

Signed-off-by: Wincent Colaiuta <win@wincent.com>