Capistrano: add 'migrate:test', 'migrate:development' etc tasks
This allows for a more conservative deployment pattern. Instead of
running migrations on all databases (including on the production
database) and _then_ running the specs, we can now run them on the
test database first, run the specs, and _then_ run them on the
other databases.

Note that to actually use the new capfile we need a patched version
of Capistrano. A patch has been submitted for inclusion upstream; see:

  http://capistrano.lighthouseapp.com/projects/8716/tickets/61

Signed-off-by: Wincent Colaiuta <win@wincent.com>