Partially revert 15ba815
So it turns out I've given up hope on Capistrano 2.5.3. No sense in
hoping for my patch to be applied if I have no intention of upgrading
from 2.2.0.

The reason is some changes after 2.2.0, such as the inability to
specify remote branch names (like "origin/master"). This is because
instead of doing a "git rev-parse" on the local repo (where "cap" is
being run) it now wants to do a "git ls-remote" to hit the remote repo.
See this commit for more details on that change:

  http://github.com/jamis/capistrano/commit/ec9a7fa5

This didn't work for me as I had specified my repository path to
Capistrano in "/pub/foo/bar" instead of what looks like is expected
by the Capistrano developers ("user@machine:/pub/foo/bar").

Changing to "user@machine:/pub/foo/bar" form allows revision querying
to work but then breaks passwordless clones, which is a huge issue for
me. It seems natural to think that the clone will be slower as well,
because it will travel through the network layer (even if only on the
loopback interface) instead of just hitting the disk. Furthermore,
git clones of local repositories are very fast and take up almost no
space due to the use of hard-links; but this behaviour is lost when you
move away from the "/pub/foo/bar" format (see "man git-clone" for more
details on that).

So that's the dealbreaker right there. Could try submitting patches for
the issue (ssh into server, cd into repo, do "git rev-parse" instead
whenever repo path is in "/pub/foo/bar" form) but I have been thinking
about security problems in Capistrano and suspect its days with me are
numbered anyway.

Due to the security concerns I'll soon start working on a set of
replacement scripts to replace Capistrano entirely. See the following
ticket for info on those:

  http://rails.wincent.com/issues/1191

One benefit of switching away from Capistrano in that way is that I'll
free myself of worries about my deployment chain breaking during updates
once and for all.

Signed-off-by: Wincent Colaiuta <win@wincent.com>