Switch to cookie-based session stores
Read/write access of the MySQL-backed session store is fine in general,
and performance is steady even as large numbers of sessions accumulate
in the database.

The problem, however, is that pruning the database is a very slow
operation that requires extended downtime. A couple of recent tests show
that takes between 30 minutes and 2 hours to purge old sessions
accumulated after a period ranging from about 1 month to just over 2
months. That is a _lot_ of downtime.

One remedy would be to reduce the amount of time sessions are allowed to
build up before being purged, perhaps to one week or less, but then
users would lose the convenience of persistent logins.

The solution being trialled here is switching to the cookie-based
session store. A security audit has been conducted and nothing sensitive
is stored in the session, so the only real drawbacks are the unwanted
transparency of the session data and the network overhead.

I would like to see the session data encrypted rather than just signed.
This would fix the transparency problem.

I'd also like salt to be used when producing the digest. Failure to use
salt really is a fundamental error that _all_ encryption wisdom advises
against. (Even though brute-forcing the secret key is unlikely, even
without salt, it's still made unnecessarily easy due to the lack of
salt.)

My login system has never relied on the session, thankfully, so I can
easily trial this alternative without worrying too much about breaking
things or compromising security.

We can easily go back to the old store by reverting this commit (I'll
be leaving the db itself intact). We can also consider trialling other
session storage options, depending on how the trial goes.

See:

  http://rails.wincent.com/issues/1142

Signed-off-by: Wincent Colaiuta <win@wincent.com>