Improve 422.html error message
Rails abuses the 422 HTTP status code for a couple of things; one is
failed validations in ActiveResource requests, and the other, at issue
here, is when its forgery protection mechanism kicks in and raises an
InvalidAuthenticityToken exception. (Both of these are "abuse" because
422 is actually only relevant to WebDAV; see RFC 4918 for details.)

So given that we know why a user might really see this, give them a
better explanation of what happened as well as some helpful advice to
fix the problem.

I learnt about this will playing with the sessions table for ticket
from the database, and you then submit, you'll see the 422 page.

Unfortanately, the only solution is to not go back and resubmit; you
actually have to go back, refresh the page to get a new token, and then
resubmit. At a minimum, this requires you to copy your form input so
that you can paste it back in after refreshing the page.

I'll be tuning the session expiry carefully to make it as long as
possible and hopefully avoid too many 422s. We don't want people going
away to make a sandwich and then getting annoyed by a 422 because they
took too long.

See also:

  http://rails.wincent.com/issues/1142

Signed-off-by: Wincent Colaiuta <win@wincent.com>