Revert "Switch to cookie-based session stores"
The cookie-backed experiment has now concluded. In itself it was
successful enough but in the analysis of what's getting stored in the
session, and the reminder that _all_ login information is stored in
entirely separate cookies, I realized that the nature of my session
data is such that I don't need to keep it around for long at all; not
a month, and not even a week.

For reasons already mentioned in d0230fd, we want to make the interval
as long as possible to avoid 422 errors, so one day would be ideal. If
it weren't for the 422 issue we could happily drop as low as 1 hour
because our session data really is that transient.

This reverts commit ceda020d98c1a597fb0d636be21ea1d875503581.

Signed-off-by: Wincent Colaiuta <win@wincent.com>