Rake: thin sessions table after 24 hours instead of 30 days
Testing shows that keeping lots of session data is a maintenance
impossibility, but that keeping only 24 hours is quite viable.

Furthermore, analysis of exactly _what_ is getting stored in the
session shows that the data is extremely ephemeral in nature
(things like flash messages, not even the current user is stored
in the session), so there was never any really reason to hang on
to so much session data anyway.

See:

  http://rails.wincent.com/issues/1142

Signed-off-by: Wincent Colaiuta <win@wincent.com>