Let admins set tags on issues
The "pending_tags" attribute is not accessible (to prevent users from
hacking together their own form submission and creating new tags), so
must explicitly set the pending_tags parameter when the remote user is
an admin.

Commit includes specs. Controller specs continue to be a source of
great consternation for me. Here I just want to make sure that
pending_tags are assigned only for admins, a fairly reasonable aspect
of the externally visible behaviour of the controller action, right?

So I initially wrote this using mock_model so that I could confirm
that the message was (or was not) received as appropriate, but that
requires me to stub out _all_ the message sends inside the action,
or the spec will fail with warnings like "received unexpected message
:user=" and so on.

That would be ridiculous, brittle work, as I would effectively be
replicating the internal structure of the tested method in order to
test one aspect of its internal behaviour.

So I swapped in a real model object and set an expecation on that,
which is much less work but still feels wrong.

I may be testing this at the wrong level (perhaps a story is more
appropriate), or perhaps I should forget mocks/stubs entirely here
and just let the action hit the database, _then_ inspect the
database to see what happened (that doesn't feel too clean, or fast,
either).

Signed-off-by: Wincent Colaiuta <win@wincent.com>