Update specs to handle issue edits involving pending tags
We need to pass in the pending tags explicitly or the specs will fail
(just as the app would bomb out in real life if we hit the issues#update
action without a proper set of form parameters).

So while we fix this breakage, add a spec to confirm that we
actually do set the pending tags.

Signed-off-by: Wincent Colaiuta <win@wincent.com>