acts_as_taggable: don't blow away old tags on save
This is a long-standing bug which lay dormant for a long time because
it wasn't exposed in the user interface.

Basically, if a model had tags on it and was saved the tags would be
blown away _unless_ the pending_tags attribute was set. So in practice
this was always true because up until now I had only been editing
posts and articles, and always using the posts#new, articles#new,
posts#edit and articles#edit forms (all of which set the pending_tags
line).

However, on adding tagging support to the issues model the bug was
exposed, because in addition to issues#new and issues#edit, these models
can also be edited using AJAX in-place editors, which obviously don't
set up pending_tags.

The only remaining bug in the whole taggable thing is that when
validation fails, you lose your tags. But that will have to be a fix for
another day.

See:

  http://rails.wincent.com/issues/1197

Signed-off-by: Wincent Colaiuta <win@wincent.com>