Stop validation errors from blowing away pending tags
Pending tags (tags a user has entered into a form, but which can't be
saved to the database yet because the taggable model in question hasn't
been saved yet) were previously blown away whenever there was a
validation error.

That is, you could enter "foo bar baz" in the tags field, and if a
validation error prevented you from saving your model then when the
form was re-rendered your "foo bar baz" tags would be gone and you'd
have to re-enter them.

The cause was that the "pending_tags" accessor always re-read the tags
back in from the database, which in such cases lead to an empty "tags"
line.

If we actually have "pending_tags" in our instance variable, use that
instead, only reading off the database if it is not set. Whenever we
successfully save our model and write our tags out to the database
"pending_tags" is set back to nil so the next time the form is displayed
it will be correct.

Signed-off-by: Wincent Colaiuta <win@wincent.com>