Fix discrepancies in tag count when non-public objects are tagged
Just say we have a public record tagged "foo" and a private record
tagged "foo". When an admin goes to the "/tags/foo" path he/she
will see:

  2 items tagged with foo
  1 wiki article
    my article
  1 issue
    the bug

An anonymous visitor would see:

  2 items tagged with foo
  1 wiki article
    my article
  0 issues

So there are really two problems here to fix. First of all we must
suppress the display of the empty "issues" group entirely. This is
done by only adding such a group to our list if it is non-empty
after being filtered for inaccessible items.

Secondly we want to display a count of taggings that we are going
to actually show to the user, not the count of taggings that exist
in the database but which we can't fully show.

Fixes:

  http://rails.wincent.com/issues/1199

Signed-off-by: Wincent Colaiuta <win@wincent.com>