Remove old version of Webrat plug-in
The Cucumber tests are only ever going to run on the local development
machine; this is because we need to test JavaScript and will have to do
in-browser tests with Selenium.

So there is no sense in bundling up Webrat (or Cucumber itself, or any
of its other dependencies) with the application. We continue to bundle
RSpec because we will always run the specs on the server; we just won't
run the acceptance tests.

Signed-off-by: Wincent Colaiuta <win@wincent.com>