Add class methods to Issue model for getting kind and status strings
These will be used at the controller level to conveniently query for
things like, "what does kind code 1 mean?". This is necessary for
the issue annotations feature, where we'll be working with a "changes"
hash provided by Rails and want to look at the "old" and "new" values
of the changed attributes, and display them in a nice human-readable
form.

Signed-off-by: Wincent Colaiuta <win@wincent.com>