Adding "pending_tags?" method
Convenience method so an external caller can check to see if a model
has pending tag changes that must still be written to the database.
This will be used by the issue annotations feature, which will want
to know if it has to describe changes made to the tags on the issue.

Signed-off-by: Wincent Colaiuta <win@wincent.com>