First stab at issue annotations
This is the first approximation at issue annotations, doing all of the
work inside the issues controller. Unfortanately it doesn't work in all
situations because there is at least one action which I don't actually
control (the "set_issue_summary" action, which is created dynamically
by the in-place editor field plug-in).

So in its current form it handles the AJAX-based edits of things like
the status pop-up, kind pop-up, and the public checkbox.

In order to handle changes made from the issues#edit form and also the
rogue method created by the in-place editor field will need to move this
logic into the controller (which in itself is complicated because the
model will need to know _who_ did the edit, and that kind of information
is normally only available in the controller).

Signed-off-by: Wincent Colaiuta <win@wincent.com>