Refactor issue annotations from controller to model layer
By handling annotations from inside the model we no longer have to
worry about where updates are being initiated. This makes things a
lot simpler because the different places you can initiate an update
each have quite different behaviour (we have the standard issues#edit
form, a number of AJAX checkboxes and pop-ups which go through a
handle_ajax_request method in the controller, and finally an in-place
editor field which throws a spanner in the works because it uses a
dynamically generated method that we can't easily change).

We don't bother doing annotations for new records, but we do for all
others. If nothing interesting changed then we do nothing.

The sole complication is that the model needs to know who the current
user is in order to assign ownership to the annotation. The normal way
to do this is to pass the info down from the controller layer
explicitly. Unfortunately, that is not an option here due to the
in-place editor field, which does an Issue.find of its own and doesn't
give us access to the record.

We get around this by using thread-local storage. This is a somewhat
nasty violation of encapsulation, little better than using a global
variable, but it seems to be the lesser of all evils. Rails is
currently single-threaded anyway to all intents and purposes, but
even if it weren't then this storage should be reasonably safe.

At the moment I am only storing the current user in the issues
controller, not in the application controller, seeing as there is not
yet any need for it elsewhere.

Signed-off-by: Wincent Colaiuta <win@wincent.com>