Clear current user from thread-local storage after request
It's probably a good idea to clear the current user from thread-local
storage at the end of each request, just in case it lingers on for
subsequent requests and causes unexpected results.

You can imagine a scenario in which the superuser edits an issue,
and later on an anonymous user comes along and has their request
serviced by the same thread. They create a new issue and it turns
out that the old "current user" info is still in the thread-local
storage.

In practice no ill effects would come of this, seeing as we
currently only use the current user info in the annotations
code path, and we only do annotations for non-new records, and
the only user who can take any actions that would cause an
annotation is the superuser, but it's still best to be conservative
about these things, in case the behaviour of the application
changes in the future.

Signed-off-by: Wincent Colaiuta <win@wincent.com>