Replace custom tag validator with standard validates_format_of
Previously we needed a custom validator because we had an accessor
that always returned valid values (pulled from the database), and
so invalid form input couldn't be tested.

Now we have an accessor which returns the form input and so we can
use the standard "validates_format_of" validator.

This makes the code simpler, and the only downside is that error
messages now refer to "pending tags" (the name of the attribute)
instead of "tags" (the name as shown in the form).

Signed-off-by: Wincent Colaiuta <win@wincent.com>