Prepare for switch to SSL transport
To avoid cookie capture attacks the plan is to serve the entire site
over SSL.

It's true that I could set ":secure => true" on my session key cookie,
and then require all logged-in users to use SSL while leaving anonymous
users connecting via normal HTTP; this would certainly be more efficient
in terms of server resources.

If a logged-in user tries to connect to an HTTP URL then the cookie
should not be sent and they won't be treated as logged in. If they try
to do something that requires them to be logged in then they'll be asked
to log in, and the login page would be accessed via HTTPS; at that moment
the user agent should send their cookie through and they will be logged
in automatically without having to enter their credentials.

The problem with this approach is that some user agents might not respect
the "Secure" attribute on the cookie. The only way to make things
completely secure from capture attacks is to _only_ provide access via
SSL.

So for now just laying down some groundwork and will switch over to SSL
in the staging environment. If benchmarking shows that it's viable, will
do the same for the production environment; and if not, back to the old
drawing board.

Signed-off-by: Wincent Colaiuta <win@wincent.com>