Unset "current_user" variable even if an exception is thrown
We previously used a before filter to store the current user in
thread-local storage, and an after filter to clear it. If
an exception derailed the request in between the two filters
then the current user value could linger on and cause
what should have been anonymous issue annotations to become
owned by the user that was current at the time of the exception.

(Discovered while refactoring code; reproducing the failure is
a little difficult, but it appears that an exception in an
unrelated test caused one of the issue model specs to fail. This
only happened when running "spec spec", not when running "rake
spec". It's a very context-sensitive issue, but the included
spec should catch the bug if it ever regresses.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>