Use HTTPS for absolutely everything
This is an initial pass at converting all URLs in the application
to use SSL.

It is a nasty, nasty change, because it is so invasive, but it is
the minimal necessary to get all specs passing.

I will be auditing the codebase subsequently and looking at all
uses of the "url_for" method to make sure that they correctly set
":protocol" to "https".

The starting point was editing the routes so that methods like
"forums_url" would generate "https://" URLs instead of "http://"
ones.

I then modified the specs to match the new behaviour (both
generation and recognition behave differently).

Note that old-fashioned routes (made with "map.connect" cannot
accept a ":protocol" option because this breaks the "url_for"
method. To see why, you can look at the Rails source and see
that Rails destructively edits the options hash passed to
"url_for", removing the ":protocol" option, and _then_ passing
the pruned optioned hash to the routing module, where a
"No route matches" RoutingError is raised because the protocol
is missing and no routes match.

For this reason, I was forced to drop use of url_for in a couple
of places and switch to named route helpers like "topic_url".

Another necessary change was modifying "acts as sortable" to at
least use "https" URLs if the current request is over SSL; this
is once again caused by problems with "url_for".

Still to be fixed: at least in the local development environment,
generated URLs in outgoing mails are wrong because they combine
an "https://" protocol string with the non-SSL port number of the
mongrel instance (3000). This may turn out to not be a problem
in the production environment, where no port numbers are actually
used.

Another step still to be undertaken is to use "_url" methods
instead of "_path" ones; without this step most of the URLs
generated by the app will inherit the protocol of the parent
page. (In other words, if a visitor guesses the mongrel port and
connects directly without SSL, then he or she will mostly see
"http://" for internal links; this is not what we want: we want
the user to be directed to the "https://" links as soon as
possible).

Signed-off-by: Wincent Colaiuta <win@wincent.com>