Fix redirection bug in articles controller
Previously could only redirect to "http://" URLs; now "https://" URLs
work as well.

I tried to come up with a regression spec for this but RSpec is still
horribly broken when trying to do controller tests for controllers
set up using the ":as" option in config/routes.rb; I simply don't know
the magical incantation, if any exists, to actually make RSpec hit the
controller and the action to be tested.

Signed-off-by: Wincent Colaiuta <win@wincent.com>