Fix spec breakage caused by b39c68c
Another "nasty" change required in the move to SSL; many specs broke
because of the way Rails sets up fake request objects for testing.
These ActionController::TestRequest instances will basically always
return false for "ssl?" and thanks to the before filter introduced
in b39c68c, that means that we'll always redirect and this breaks
many specs.

As you can see the fix is "nasty" because it requires two ugly
changes:

- first, a monkey patch to the TestRequest class to make it always
  return true for "ssl?"; it's not possible to actually have it
  inspect itself to see whether it should return true because the
  instances are wildly inconsistent (sometimes ":protocol =>
  'https'" is set in the parameters instance variable, sometimes in
  @path_parameters, sometimes in @symbolized_path_parameters, and
  sometimes not at all).

- second, change our nested_comments_path method into
  nested_comments_url; this is a "nasty" change because it requires
  us to change the application for what is essentially a problem in
  the tests.

Signed-off-by: Wincent Colaiuta <win@wincent.com>