Use "_url" methods instead of "_path" methods
For consistency we use "_url" methods pretty much everywhere now,
apart from a few specific places where we _must_ use paths (cache
sweepers, for example).

In addition to consistency, this is also another layer in a
"defense in depth" strategy designed to prevent access to resources
over insecure lengths. If nginx fails to redirect HTTP requests to
HTTPS, or if our app-level before filter fails to redirect in the
case of direct access to our mongrels, at least the URLs in the
generated HTML will point to HTTPS resources.

This will also be useful if I ever decide to partially roll back
some of these recent changes (for example, use SSL only for some
resources or for some users). In that case need to use "_url"
methods in any case rather than "_path" ones, for those occasions
when you move from HTTPS to HTTPS and vice versa.

Signed-off-by: Wincent Colaiuta <win@wincent.com>