capistrano: add "unlock" and "lockdown" tasks
The problem: each Rails app runs as a specific, unprivileged user
for greater security, but said users must have login shells so
that Capistrano can actually do deployments. It is desirable for
these users to have fake shells such as "/sbin/nologin" to
partly contain the damage in the event of a web application
vulnerability.

I looked into many different ways around this problem and they
all required some kind of ghastly kludge (such as entirely
replacing Capistrano, or making some other undesirable security
compromise). For full details of all the measures considered
see:

  https://wincent.com/issues/1191

In the end, the "lesser of all kludges" turned out to be the one
implemented in this commit. Basically, our app users have
"/sbin/nologin" as their shell 99% of the time, and we temporarily
swap in "/bin/sh" immediately prior to deploying, swapping it
back out afterwards.

This is achieved by means of two new Capistrano tasks,
"deploy:unlock" (to relax security) and "deploy:lockdown" (to
tighten it).

There is still a small window of opportunity in which the app
users have a login shell, but we've at least eliminated the
hideous leaking sieve that was previously in place 24 hours a
day, 7 days a week.

Signed-off-by: Wincent Colaiuta <win@wincent.com>