Horrible kludge to make Cucumber stories work again
Since moving to an SSL-only policy the Cucumber stories are
all broken. The hideous hack implemented in spec/spec_helper.rb
to stop unwanted redirects from kicking in (see commit b2bde10)
won't work for Cucumber for some unknown reason (either it
isn't using ActionController::TestRequest under the hood, or it
is overwriting my override of the "ssl?" method).

Seems that the only way to fix the breakage is a horrible runtime
test; this is horrid because code should _not_ conditionally
modify its behaviour just because it detects that the test suite
is being run.

We at least confine the ugliness to the end of the "if" statement,
so that the conditional behaviour is tested for only in the
minority case where access is _not_ via HTTPS.

See also:

  http://rubyforge.org/pipermail/rspec-users/2009-February/012347.html

Signed-off-by: Wincent Colaiuta <win@wincent.com>