rspec: change comment for "ssl?" override
The horrible kludge introduced in 8a41e79 means that _this_
horrible kludge is (mostly) no longer necessary. Unfortunately
without it we fail two specs in the suite, namely:

1) 'SessionsController redirecting after login should redirect
    to if original uri supplied via params' FAILED
expected redirect to "https://test.host/comments",
     got redirect to "http://test.host/comments"
./spec/controllers/sessions_controller_spec.rb:99:

2) 'SessionsController redirecting after login should redirect
    to if original uri supplied via session' FAILED
expected redirect to "https://test.host/comments",
     got redirect to "http://test.host/comments"
./spec/controllers/sessions_controller_spec.rb:94:

Seems that the controller does a "redirect_to original_uri",
and given that the original_uri is just "/comments" (no protocol
or hostname), Rails decides to construct a URL based on what the
ActionController::TestController object thinks is correct;
namely, that the host is "test.host" and the protocol is "http"
because "ssl?" returns false.

So regrettably, we either need to patch these specs or just keep
the "ssl?" override hack in place; we chose to do the latter as
it seems the more maintainable option in the long run.

Signed-off-by: Wincent Colaiuta <win@wincent.com>