Add (broken) prototype for "expire_all" functionality to sweepers
This really should be more than a prototype (it should be a fully
working implementation) but there is some "magic" going on in Rails
that prevents it from working.

Basically, the "expire_page" message is sent and no exception is
raised, but exactly _where_ it is sent is a complete mystery as
logging reveals that the "expire_page" methods defined by Rails
are never called.

So looks like I will have to manually expire these pages using
"File.delete" and friends. Unfortunately I cannot even call
"page_cache_directory" to dynamically get the cache directory at
runtime, because I don't know what object responds to that message;
the application controller might, apparently, but I can't/shouldn't
try accessing that from a sweeper:

  ActionController::Base.new.page_cache_directory
  => "/Users/wincent/trabajo/unversioned/wincent.com/src/public"

Signed-off-by: Wincent Colaiuta <win@wincent.com>