Fix on-demand cache expiration methods in sweepers
So I worked out why the "expire_page" method doesn't work
in my on-demand cache expiration methods: turns out that
Rails defines some method_missing "magic" in
ActionController::Caching::Sweeper to forward those
messages on to the appropriate controller. If there is no
such controller (and there won't be unless the sweeper is
set up and called from a controller) then the message just
gets dropped on the floor.

So we have to manually construct the path to each cached
page and do a "File.delete" on it. Unfortunately even
this is made uglier than I would like; the "page_cache_path"
method itself is private and so we need to hit it with
"send", and our routing helpers also don't work (presumably
because they are actually private methods) unless we use
"send" for those too.

Signed-off-by: Wincent Colaiuta <win@wincent.com>