Update to RSpec 1.11.99.7 (Rspec 1.2 RC1), unvendoring it
Note that this commit effectively _unvendors_ RSpec seeing
as "rake gems:unpack" is currently broken. But updating is
still desirable to help prepare for when Rails 2.3 goes
final (which itself will probably require RSpec 1.2).

Lots of routing specs were broken by this update, but the
fixes are included in this commit.

Signed-off-by: Wincent Colaiuta <win@wincent.com>