Explain why we don't "config.gem" Cucumber
Because it's only used in local testing right now (because we'll
be doing stuff like in-browser testing).

Will eventually need to look at separating out the Cucumber suite
into two parts: a part that does in-browser testing and only runs
on the local development machine, and a part that does normal
Webrat testing and runs on the server too. It might also be
possible to look at one of those "browser simulator" packages so
that we could do the AJAX tests even on the server. See:

  http://www.mozilla.org/rhino/ (JavaScript engine written in Java)
  http://celerity.rubyforge.org/ (JRuby wrapper for HtmlUnit)
  http://htmlunit.sourceforge.net/ (headless browser with JS support)

Signed-off-by: Wincent Colaiuta <win@wincent.com>