Add "Message" model
This model is used to capture debugging and tracking
information about both incoming and outgoing email
messages.

Outgoing messages, for example, are recorded here so
that when a user clicks "Reply" the "In-Reply-To"
header can (hopefully) be matched up with the
"Message-ID" of the outgoing message and basic
"threading" in the issue tracker can be achieved.

Incoming messages are either assigned to an existing
ticket in this way, or a new ticket is created.

Signed-off-by: Wincent Colaiuta <win@wincent.com>