Move custom url options out of mailers
These can be set just once in the "config/environment.rb"
file; there is no need to set them in all of the mailers.

Note that if a mailer ever uses the "url_for" method then
it will need to pass in ":only_path => false" because
otherwise a "relative" URL (eg "/issues/1234") will be
generated (this happens by default when no explicit ":host"
option is passed in). For now this isn't an issue as no
mailer uses url_for; they all use custom route helpers.

Signed-off-by: Wincent Colaiuta <win@wincent.com>